FIX: VisualStudio compiler warnings
git-svn-id: https://www.nitrc.org/svn/cmtk/trunk@161 42a5c34f-2066-0410-bec5-ba365beb4995

+++ core/libs/Registration/cmtkCongealingFunctional.cxx
@@ -403,7 +403,7 @@ CongealingFunctional<TXform,THistogramBinType>
 
     for ( size_t smpl = sampleFrom; smpl < sampleTo; ++smpl )
       {
-      float sum = 0, sumsq = 0;
+      double sum = 0, sumsq = 0;
       unsigned int count = 0;
 
       if ( ThisConst->m_UseTemplateData )
